chore: refactored facing diff test, made the calculators package private so they are not used outside of princess
